requirements.txt
# Core LangChain
langchain
langchain-community
langchain-core
langchain-text-splitters

# LLM Providers
langchain-ollama
langchain-anthropic

# Vector Store
langchain-chroma
chromadb

# Web UI
streamlit

# Utilities
python-dotenv
faiss-cpu
